Colorize std::stringstream if 'colorize' is passed
So far termcolor aimed to colorize only those streams that are bound to
TTY (terminal). However, it's not always the case and sometimes we need
to colorize (i.e. write ANSI escape sequences) text even in plain
stringstreams, so when we print them to TTY - it will be colorized.

For this case two new manipulators were added:

 * colorize
 * nocolorize

which are enable and disable colorization respectively.

Usage:

 std::stringstream ss;
 ss << termcolor::colorize << termcolor::blue << "Hej!";

Fixes #5