Add CMake usage examples
CMake is a popular tool among C++ developers, however, not trivial and
covered with poor documentation. That's why it would be nice to provide
a couple of good examples of how one can use termcolor library by
utilizing CMake features. Besides, these examples could be used on CI to
ensure usecases are not broken.