Don't error when color/on_color is used on Windows
color<>/on_color<> are new stream manipulators that can be used to use
any color from 256 color table. This table is not supported on Windows
but we don't want to trigger compilation errors in this case. Instead,
we want to be consistent with other manipulators which are not available
on windows and simply fallback to no-op.