Share colorization flag between translation units
Termcolor is used to provide `colorize`/`nocolorize` stream manipulators
to colorize streams other than ones ties to TTY. Such streams are marked
by termcolor as capable to carry ASCII color codes. However, due to the
fact that the static variables are initialized per translation units, it
turns out that colorized stream returned by one translation unit to
another are not recognized as color capable. This patch makes things
right and ensures that the flag we use to store the marker is shared and
the same between translation units.

Fixes #54