Added wchar_t support (#58)
Add wchar_t support

wchar_t - type for wide character representation (see wide strings). Required
to be large enough to represent any supported character code point (32 bits on
systems that support Unicode. A notable exception is Windows, where wchar_t is
16 bits and holds UTF-16 code units) It has the same size, signedness, and
alignment as one of the integer types, but is a distinct type. 

There are C++ stream objects that are pointing to stdout/stderr and that are
dealing with wide characters: wcout, wclog, wcerr. This patch essentially
brings their support.